requirements.txt
streamlit
streamlit_webrtc
paho-mqtt==1.6.1
opencv-python
mediapipe==0.8.9.1